refactor: Add DicController
